# Add some precompiles to help loading time